Comitando parte CRUD do tutorial

+++ UniversidadeDeContoso/Controllers/EstudantesController.cs
@@ -31,7 +31,10 @@ namespace UniversidadeDeContoso.Controllers
             }
 
             var estudante = await _context.Estudantes
-                .FirstOrDefaultAsync(m => m.ID == id);
+                .Include(e => e.Materias)
+                .ThenInclude(m => m.Curso)
+                .AsNoTracking()
+                .FirstOrDefaultAsync(e => e.ID == id);
 
             if (estudante == null)
             {
@@ -52,16 +55,24 @@ namespace UniversidadeDeContoso.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ID,SobreNome,Nome,DataDeMatricula")] Estudante estudante)
+        public async Task<IActionResult> Create([Bind("SobreNome,Nome,DataDeMatricula")] Estudante estudante)
         {
-            if (ModelState.IsValid)
+            try
             {
-                _context.Add(estudante);
+                if (ModelState.IsValid)
+                {
+                    _context.Add(estudante);
 
-                await _context.SaveChangesAsync();
+                    await _context.SaveChangesAsync();
 
-                return RedirectToAction(nameof(Index));
+                    return RedirectToAction(nameof(Index));
+                }
             }
+            catch (DbUpdateException error)
+            {
+                ModelState.AddModelError("", "Não foi possivel salvar os dados." +
+                    $"Tente de novo, e se o problema persistir contate o administrador do sistema, error:{error}");
+            } 
             return View(estudante);
         }
 
@@ -84,52 +95,48 @@ namespace UniversidadeDeContoso.Controllers
         // POST: Estudantes/Edit/5
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
+        [HttpPost, ActionName("Edit")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ID,SobreNome,Nome,DataDeMatricula")] Estudante estudante)
+        public async Task<IActionResult> EditPost(int? id)
         {
-            if (id != estudante.ID)
-            {
+            if (id == null)
                 return NotFound();
-            }
 
-            if (ModelState.IsValid)
+            var estudandoParaEdicao = await _context.Estudantes.FirstOrDefaultAsync(e => e.ID == id);
+
+            if (await TryUpdateModelAsync(estudandoParaEdicao, "", e => e.Nome, e => e.SobreNome, e => e.DataDeMatricula))
             {
                 try
                 {
-                    _context.Update(estudante);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (DbUpdateException error)
                 {
-                    if (!EstudanteExists(estudante.ID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    ModelState.AddModelError("", "Não foi possivel salvar os dados." +
+                        $"Tente de novo, e se o problema persistir contate o administrador do sistema, error:{error}");
                 }
-                return RedirectToAction(nameof(Index));
             }
-            return View(estudante);
+            return View(estudandoParaEdicao);
         }
 
         // GET: Estudantes/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? mudancasAoSalvar = false)
         {
             if (id == null)
-            {
                 return NotFound();
-            }
 
             var estudante = await _context.Estudantes
-                .FirstOrDefaultAsync(m => m.ID == id);
+                .AsNoTracking()
+                .FirstOrDefaultAsync(e => e.ID == id);
+
             if (estudante == null)
-            {
                 return NotFound();
-            }
+
+            if (mudancasAoSalvar.GetValueOrDefault())
+                ViewData["MensagemDeErro"] =
+                    "Exclusão do estudante falhou. Tente novamente, e se o problema persistir " +
+                    "contate o administrador do seu sistema.";
 
             return View(estudante);
         }
@@ -137,12 +144,25 @@ namespace UniversidadeDeContoso.Controllers
         // POST: Estudantes/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> DeleteConfirmado(int id)
         {
-            var estudante = await _context.Estudantes.FindAsync(id);
-            _context.Estudantes.Remove(estudante);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            var estudante = await _context.Estudantes
+                .FindAsync(id);
+
+            if (estudante == null)
+                return RedirectToAction(nameof(Index));
+
+            try
+            {
+                _context.Estudantes.Remove(estudante);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException /* ex */)
+            {
+                //Log the error (uncomment ex variable name and write a log.)
+                return RedirectToAction(nameof(Delete), new { id, saveChangesError = true });
+            }
         }
 
         private bool EstudanteExists(int id)
